fix: dicimal to float

+++ app/repository/wallets.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 
 from sqlalchemy import select
-from sqlalchemy.ext.asyncio import AsyncSession
+
 from app.database import SessionLocal
 from app.models import Wallet
 
@@ -12,7 +12,7 @@ async def wallet_already_exist(wallet_name: str) -> bool:
         result = await db.execute(query)
         return result.scalar_one_or_none() is not None
     
-async def add_income(wallet_name: str, amount: float) -> Wallet | None:
+async def add_income(wallet_name: str, amount: Decimal) -> Wallet | None:
     async with SessionLocal() as db:
         query = select(Wallet).filter(Wallet.name == wallet_name)
         result = await db.execute(query)
@@ -30,7 +30,7 @@ async def get_wallet_balance_by_name(wallet_name: str) -> Wallet | None:
         result = await db.execute(query)
         return result.scalar_one_or_none()
 
-async def add_expense(wallet_name: str, amount: float) -> Wallet | None:
+async def add_expense(wallet_name: str, amount: Decimal) -> Wallet | None:
     async with SessionLocal() as db:
         query = select(Wallet).filter(Wallet.name == wallet_name)
         result = await db.execute(query)
@@ -48,7 +48,7 @@ async def get_all_wallets() -> list[Wallet]:
         return list(result.scalars().all())
 
 
-async def create_wallet(wallet_name: str, amount: float) -> Wallet:
+async def create_wallet(wallet_name: str, amount: Decimal) -> Wallet:
     async with SessionLocal() as db:
         new_wallet = Wallet(name=wallet_name, balance=amount)
         db.add(new_wallet)
