+++ app/schemas.py
@@ -1,13 +1,15 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 
 class Operation(BaseModel):
     wallet_name: str = Field(..., max_length=50)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=200)
 
     @field_validator('amount')
-    def validate_amount(cls, value: float) -> float:
+    def validate_amount(cls, value: Decimal) -> Decimal:
         if value <= 0:
             raise ValueError('Сумма должна быть положительной.')
         return value
@@ -24,7 +26,7 @@ class Operation(BaseModel):
 
 class CreateWallet(BaseModel):
     name: str = Field(..., max_length=50)
-    initial_balance: float = 0.0
+    initial_balance: Decimal = 0.0
     
     @field_validator('name')
     def validate_name(cls, value: str) -> str:
@@ -35,7 +37,7 @@ class CreateWallet(BaseModel):
         return value
 
     @field_validator('initial_balance')
-    def validate_initial_balance(cls, value: float) -> float:
+    def validate_initial_balance(cls, value: Decimal) -> Decimal:
         if value < 0:
             raise ValueError('Сумма должна быть положительной.')
         return value
